fix(leagueMember):float -> integer rate

+++ src/main/java/com/moayong/api/domain/leaguemember/dto/response/MatchInfoResponse.java
@@ -10,7 +10,7 @@ public record MatchInfoResponse (
         Integer totalScore,
         Integer goalAmount,
         Integer rank,
-        Float rate,
+        Integer rate,
         Integer nextLevel,
         PromotionStatus promotionStatus
 ){
